fix(execution): fractional backend capability negotiation + sell lifecycle
Addresses Codex CHANGES_REQUESTED on PR #153.

#1 Live vs validation backends no longer diverge into a zero-share Fill.
   Added BACKEND CAPABILITY NEGOTIATION:
   - ExecutionBackend.supports_fractional (default False; whole-share only).
   - FakeBackend / SimBackend take allow_fractional=True and MODEL fractional
     quantities (float positions, no int() floor) so the readonly/shadow/sim
     path validates the exact behaviour enabled live.
   - New types.resolve_fill_quantity() is the single negotiation point: a
     fractional-capable backend keeps the float; a whole-share backend returns
     the integral count ONLY when the request is whole, else FAILS FAST. No
     code path can floor a sub-1-share order to a zero-share fill.
   - LeanBackend stays whole-share only and fails fast on a fractional intent
     (broker/LEAN fractional contract is renquant-execution #19).
   - PrepareExecutionTask fails fast at the top of the bar when
     execution.fractional_shares.enabled is on but the backend can't model it.

#2 Full fractional position lifecycle (BUY -> upsert -> partial/full sell ->
   prune) no longer truncates or strands residual state:
   - Backend SELL full-liquidate sells the ENTIRE float (never int()-floors a
     sub-1-share holding to 0); fp dust is clamped to exactly 0 so
     PruneFullExitsTask reaps the holding.
   - ExecuteExitsTask preserves a FLOAT partial-sell quantity on a
     fractional-capable backend (was int(float(sig.quantity))).
   - TrimHeldTask trims to the exact Kelly target as a float under fractional
     (dust-guarded), instead of int()-flooring the partial sell.
   - rotation / joint sell-leg proceeds and the QP close/sell reads use the
     float held quantity so a fractional sell-leg is not truncated to 0.
   QP/benchmark BUY sizing stays whole-share (not a fractional emit path);
   any fractional position they encounter is always fully closable.

#4 Type/config discipline:
   - OrderIntent BUY/partial-SELL shares reject bool and string-likes
     (require a finite positive real), matching the Fill guard.
   - fractional_sizing_cfg fails CLOSED: enabled must be a genuine bool (the
     YAML string "false"/"true" no longer flips it on); min_notional rejects
     bool and non-finite/negative values.

Tests: new tests/test_fractional_execution.py covers sub-1-share fractional
Fill (not zero) on a capable backend; whole-share/LEAN fail-fast (no silent
zero); fractional buy->partial->full-sell lifecycle with cash conservation and
clean holding reap; pipeline negotiation guard; and the type/config guards.
Full pipeline suite green (1029 passed, 8 skipped); ruff + doctor clean.

#3 (broker GTC/stop fractional contract) is cross-repo in renquant-execution
#19 and remains the gate to LIVE enablement; this PR keeps fractional default
OFF and pairs with that work — it does not enable live on its own.

+++ src/renquant_pipeline/kernel/pipeline/task_joint_actions.py
@@ -778,7 +778,9 @@ class JointActionTask(Task):
             # mark-to-market value net of fees/slippage on the sell.
             sell_proceeds = 0.0
             if a.kind == "rotate":
-                h_shares = int(getattr(a.held_obj, "shares", 0) or 0)
+                # Fractional-share lifecycle (#153): read held qty as a FLOAT so
+                # a sub-1-share sell-leg isn't truncated to 0 in proceeds sizing.
+                h_shares = float(getattr(a.held_obj, "shares", 0.0) or 0.0)
                 h_price  = float(ctx.prices.get(a.held_ticker, 0.0) or 0.0)
                 if h_shares > 0 and math.isfinite(h_price) and h_price > 0:
                     sell_proceeds = h_shares * h_price * (1.0 - fee_pct - slip_pct)
